feat(store): Step 7 — drop the JSONB legacy + retire the backfill subsystem
The relational migration (cut-sets a–e, Steps 2–6) is complete, so the JSONB
document and the backfill that copied it into the relational tables are dead.
Step 7 removes both.

- Delete the whole lib/db/backfill/ engine (engine/gate/markers/normalize/roots/
  types + all five cut-sets) and their tests; nothing copies JSONB→relational
  anymore.
- Remove ensureStoreReady() and its 10 call sites (auth, membership, the two
  GraphQL paths, the github-webhook + agent-bootstrap routes, instrumentation
  boot). It was purely the backfill gate. Delete lib/store.ts; relocate its sole
  remaining live concern — the "DEPLO_DATABASE_URL required" boot fail-fast — into
  lib/db/pg.ts as a module-load guard.
- Delete the JSONB document layer: document-store.ts (loadDocument/saveDocument),
  lib/migrate.ts (the forward-migrate) + its test, lib/seed.ts (buildSeed). Remove
  the now-unreferenced DeploData interface from lib/types.ts.
- schema/legacy.ts held BOTH deplo_state (drop) AND the LIVE scheduler_lease mutex
  (keep): relocate schedulerLease to schema/scheduler.ts, drop deploState, delete
  legacy.ts, update the schema.ts aggregate.
- Migration 0004 drops only deplo_state + store_migration; scheduler_lease is
  untouched. Remove the storeMigration table declaration. schema.test.ts drops the
  two from the expected set (proving 0004 removed them) and keeps scheduler_lease.
- Update stale doc comments that referenced the removed files (the globalThis
  Symbol.for pattern now cites lib/db/client.ts; the row-assembler seams drop their
  backfill half; the timestamp-parser notes drop document-store.ts).

KEPT: the read-time per-entity normalizers (normalize-project.ts) and the live
row-assembler seams — these are the live read↔write path, not the JSONB migrate.

Full suite 452 pass / 0 fail (−48: the deleted backfill/migrate test files covered
the removed machinery; live read paths keep their per-collection tests). tsc clean,
eslint clean (pre-existing warnings only), db:generate reports no drift.
