+++ app/api/agent/bootstrap/route.ts
@@ -1,4 +1,3 @@
-import { ensureStoreReady } from "@/lib/store";
 import { completeBootstrap } from "@/lib/data/servers";
 import { signResponse, BootstrapError } from "@/lib/agent/bootstrap";
 
@@ -7,8 +6,8 @@ import { signResponse, BootstrapError } from "@/lib/agent/bootstrap";
  * caller is a brand-new agent that has no session and no mTLS identity yet — its
  * trust comes entirely from the single-use bootstrap token (operator
  * authorisation) + the CSR's proof-of-possession, NOT from a logged-in user. So,
- * like the GitHub webhook, it bypasses getCurrentUser() and must run the backfill
- * gate (ensureStoreReady) itself before reading the relational `servers` table.
+ * like the GitHub webhook, it bypasses getCurrentUser() and reads the relational
+ * `servers` table directly.
  *
  * Flow:
  *   1. The agent has already authenticated US (P2/P3): over HTTPS it pinned our
@@ -24,8 +23,6 @@ import { signResponse, BootstrapError } from "@/lib/agent/bootstrap";
  *      returned CA — the value that anchors all future mTLS.
  */
 export async function POST(request: Request) {
-  await ensureStoreReady();
-
   let body: {
     token?: unknown;
     csrPem?: unknown;
