+++ lib/db/schema.ts
@@ -7,14 +7,14 @@
  *  - `schema/auth.ts`          — the Better Auth tables (user/session/account/
  *                                verification), owned by Better Auth's Drizzle
  *                                adapter.
- *  - `schema/legacy.ts`        — `deplo_state` (the single-row JSONB control-plane
- *                                document) + `scheduler_lease` (the backup-
- *                                scheduler mutex). Rollback artifacts kept through
- *                                the migration.
- *  - `schema/control-plane.ts` — the full relational normalization of `DeploData`
- *                                (Step 1). NO JSONB columns; every nested object
- *                                is a 1-to-1 child table and every list an ordered
- *                                child / junction.
+ *  - `schema/scheduler.ts`     — `scheduler_lease`, the cross-process backup-
+ *                                scheduler mutex (a live table; the legacy
+ *                                `deplo_state` JSONB document was dropped in
+ *                                PLAN Step 7).
+ *  - `schema/control-plane.ts` — the full relational normalization of the former
+ *                                JSONB control plane (Step 1). NO JSONB columns;
+ *                                every nested object is a 1-to-1 child table and
+ *                                every list an ordered child / junction.
  *
  * This file re-exports every table and an aggregated `schema` object, so
  * `drizzle.config.ts` (`schema: "./lib/db/schema.ts"`) and `lib/db/client.ts`
@@ -22,11 +22,11 @@
  */
 
 export { user, session, account, verification } from "./schema/auth";
-export { deploState, schedulerLease } from "./schema/legacy";
+export { schedulerLease } from "./schema/scheduler";
 export * from "./schema/control-plane";
 
 import { user, session, account, verification } from "./schema/auth";
-import { deploState, schedulerLease } from "./schema/legacy";
+import { schedulerLease } from "./schema/scheduler";
 import * as controlPlane from "./schema/control-plane";
 
 export const schema = {
@@ -34,7 +34,6 @@ export const schema = {
   session,
   account,
   verification,
-  deploState,
   schedulerLease,
   ...controlPlane,
 };
